renames station_stream to station
