+++ lib/core/providers/local_stations_state.dart
@@ -1,6 +1,6 @@
 import 'package:radio_app/core/abstract_classes/order.dart';
 import 'package:radio_app/core/providers/stations_state.dart';
-import 'package:radio_app/model/station_stream/stations_filter.dart';
+import 'package:radio_app/model/station/stations_filter.dart';
 
 class LocalStationsState extends StationsState {
   // final StationStreamFilter _filter = StationStreamFilter(countrycode: "NL", order: Order.clickCount);
