+++ lib/model/station/station.dart
